Moved files and started space spec
