+++ app/models/board.rb
