+++ app/models/ship.rb
