+++ app/models/space.rb
