+++ spec/models/space_spec.rb
@@ -1,0 +1,11 @@
+require 'rails_helper'
+
+describe Space do
+  describe "instance methods" do
+    let(:attacked_space) { create(:space, contents: ) }
+    let(:hit_space) { create(:space, status: "Hit") }
+    describe "#not_attacked?" do
+      it "it evaluates status"
+    end
+  end
+end
